dtls: use uint16_t for mseq (handshake sequence number)
Using an int makes calculation much easier.

+++ state.h
@@ -50,7 +50,7 @@ typedef enum {
 } dtls_state_t;
 
 typedef struct {
-  uint24 mseq;		     /**< handshake message sequence number counter */
+  uint16_t mseq;	     /**< handshake message sequence number counter */
 
   /** pending config that is updated during handshake */
   /* FIXME: dtls_security_parameters_t pending_config; */
